#115 [mod] : change news id type int to long

+++ core/model/src/main/java/com/teamwable/model/news/NewsInfoModel.kt
@@ -5,7 +5,7 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class NewsInfoModel(
-    val newsId: Int,
+    val newsId: Long,
     val newsTitle: String,
     val newsText: String,
     val newsImage: String?,
